Adds auto detecting features to the library.
This uses UPNP to send out a SSDP discovery packet to locate a TV on 
the local network. It collects data from the returned web address so 
the user has the ability to select which TV to control if more then 1 
TV is on the network.

All of the variables that were used previously are still in place so 
those variables can be hard coded bypassing the modifications I have 
made.